🔨 Configure Java symlink and update dependencies
Set up Java symlink for java_home utility.

This commit configures a symlink to the latest OpenJDK installation
in /Library/Java/JavaVirtualMachines/openjdk.jdk, which is used by
the java_home utility. The location of the OpenJDK installation is
determined by the architecture (arm64 or x86_64) and the
corresponding Homebrew prefix (/opt/homebrew or /usr/local).

Also, the tap for koekeishiya/formulae is explicitly stated when
installing skhd and yabai.

- Add Java symlink setup in macos/brew.sh
- Update Brewfile to use openjdk formula without version specifier.
- Remove redundant JAVA_HOME and PATH settings from sh/macos.sh